%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  R/Covariance.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{clone.Covariance}
\alias{clone.Covariance}
\alias{Covariance.clone}
\alias{clone.Covariance}
\alias{clone,Covariance-method}


\title{Deep-clones a Dataset}

\description{
  Deep-clones a Dataset.  (Clones LazyMatrix objects too.)
}

\usage{\method{clone}{Covariance}(this, ...)}

\arguments{
  \item{this}{The Covariance object to clone.}
  \item{...}{Not used.}
}

\author{Charles R. Hogg III}

\value{A deep clone of the Covariance object.}

\seealso{
  \code{\link[R.oo]{clone.Object}}
  For more information see \code{\link{Covariance}}.
}
\keyword{internal}
\keyword{methods}
