\name{clone.Model}
\alias{clone.Model}
\title{Deep-clone a Model}
\usage{
  \method{clone}{Model} (this, ...)
}
\arguments{
  \item{this}{The Model object to clone.}

  \item{...}{Not used.}
}
\value{
  A deep clone of the Model object.
}
\description{
  Deep-clone a Model (i.e., clones Covariance objects too).
}
\seealso{
  \code{\link{Model}}
}

