\name{Dataset$id}
\alias{Dataset$id}
\alias{getId}
\alias{getId.Dataset}
\alias{id}
\alias{setId}
\alias{setId.Dataset}
\title{The id string for this Dataset}
\arguments{
  \item{...}{Not used.}
}
\value{
  A string representing the ID for this Dataset.
}
\description{
  Each Dataset has an id string to give a hint about what
  it is.  These are not checked for uniqueness.
}
\details{
  Usually accesed using \code{this$id} syntax.
}
\seealso{
  \code{\link{Dataset}}
}

