\name{DeleteRows.Dataset}
\alias{DeleteRows}
\alias{DeleteRows.Dataset}
\title{Delete datapoints}
\arguments{
  \item{indices}{The indices of the datapoints to be
  removed.}

  \item{...}{Not used.}
}
\value{
  \code{this}, but used for its side-effect of deleting
  rows.
}
\description{
  Removes the indicated datapoints from the Dataset.
}
\seealso{
  \code{\link{Dataset}}
}

