\name{FixConstParam.Covariance}
\alias{FixConstParam.Covariance}
\title{Set a parameter to a constant value.}
\arguments{
  \item{p.name}{The (undecorated) name of the parameter to
  change.}

  \item{p.value}{The new constant value for the named
  parameter.}

  \item{...}{Not used.}
}
\description{
  Sets the lower and upper bounds equal for a given
  parameter, effectively fixing that parameter as constant.
}
\seealso{
  \code{\link{getParams.Covariance}}

  \code{\link{Covariance}}
}

