\name{getContributionIds.Model}
\alias{Model$contributionIds}
\alias{getContributionIds}
\alias{getContributionIds.Model}
\title{Id's of each contributing Covariance}
\arguments{
  \item{this}{The Model object.}

  \item{...}{Not used.}
}
\value{
  The id's for this Model's contributing Covariance
  objects.
}
\description{
  Recall that a Model consists of a collection of
  Covariance objects.  This function gives their ID's.
}
\seealso{
  \code{\link{Model}}
}

