\name{getD.Dataset}
\alias{d}
\alias{getD}
\alias{getD.Dataset}
\title{The dimensionality of the Dataset}
\arguments{
  \item{...}{Not used.}
}
\value{
  An integer representing the dimensionality of the
  independent variable.  (i.e., 1 for curves, 2 for
  surfaces, etc.)
}
\description{
  The dimensionality of the independent variable in this
  Dataset. i.e., 1 for curves, 2 for surfaces, etc.
}
\seealso{
  \code{\link{Dataset}}
}

