\name{getDataOffset.Dataset}
\alias{dataOffset}
\alias{getDataOffset}
\alias{getDataOffset.Dataset}
\title{The offset for this dataset}
\arguments{
  \item{...}{Not used.}
}
\value{
  The datapoint offset values for the currently-selected
  quantity.
}
\description{
  The offset of the currently-selected quantity in the
  dataset.
}
\details{
  Usually accesed using \code{this$d} syntax; read-only.
}
\seealso{
  \code{\link{Dataset}}
}

