\name{getDpts.Dataset}
\alias{dpts}
\alias{getDpts}
\alias{getDpts.Dataset}
\title{Datapoint values}
\arguments{
  \item{...}{Not used.}
}
\value{
  The raw datapoint values for the currently-selected
  quantity.
}
\description{
  Values for the currently-selected quantity.
}
\details{
  Usually accesed using \code{this$dpts} syntax; read-only.
}
\seealso{
  \code{\link{Dataset}}

  \code{\link{xformedDpts}}
}

