\name{getId.Covariance}
\alias{Covariance$id}
\alias{getId.Covariance}
\alias{setId.Covariance}
\title{The ID for this Covariance}
\arguments{
  \item{id}{A character string identifying this Covariance
  object.}

  \item{...}{Not used.}
}
\value{
  The id of this Covariance object.
}
\description{
  The ID is a character string identifying this Covariance
  object.
}
\seealso{
  \code{\link{Covariance}}
}

