\name{getLogspaceNames.CovarianceSEAniso2D}
\alias{CovarianceSEAniso2D$logspaceNames}
\alias{getLogspaceNames.CovarianceSEAniso2D}
\title{Names of "scale"-type parameters}
\arguments{
  \item{...}{Not used.}
}
\value{
  Names of parameters to be optimized in logspace.
}
\description{
  A character vector of names, indicating which parameters
  are considered to be "scale" parameters.  (Read
  \dQuote{Optimization mode} section of
  \code{\link{getParams.Covariance}} to see what this
  means.)
}
\seealso{
  \code{\link{CovarianceSEAniso2D}}
}

