\name{getLogspaceNames}
\alias{getLogspaceNames}
\title{Names of "scale"-type parameters}
\arguments{
  \item{\dots}{Not used.}
}
\value{
  Names of parameters to be optimized in logspace.
}
\description{
  A character vector of names, indicating which parameters
  are considered to be \dQuote{scale} parameters.  (Read
  \dQuote{Optimization mode} section of
  \code{\link{getParams.Covariance}} to see what this
  means.)
}
\seealso{
  \code{\link{CovarianceSE}}
}

