\name{getN.Dataset}
\alias{getN}
\alias{getN.Dataset}
\alias{n}
\title{The number of datapoints}
\arguments{
  \item{...}{Not used.}
}
\value{
  (numeric) The number of datapoints in this Dataset.
}
\description{
  Gives the number of datapoints in this Dataset object.
}
\details{
  Usually accesed using \code{this$n} syntax; read-only.
}
\seealso{
  \code{\link{Dataset}}
}

