\name{getNoiseVar.Dataset}
\alias{getNoiseVar}
\alias{getNoiseVar.Dataset}
\alias{noiseVar}
\title{Noise variance for current column}
\arguments{
  \item{...}{Not used.}
}
\value{
  (numeric) The noise variance for the currently selected
  column.
}
\description{
  The noise variance for the quantity in this Dataset which
  is currently selected.
}
\details{
  Usually accesed using \code{this$noiseVar} syntax;
  read-only.
}
\seealso{
  \code{\link{Dataset}}

  \code{\link{quantity}}
}

