\name{getParamNamesPlain.CovarianceSE}
\alias{CovarianceSE$paramNamesPlain}
\alias{getParamNamesPlain.CovarianceSE}
\title{Basenames of parameters}
\arguments{
  \item{...}{Not used.}
}
\value{
  The basenames of the parameters.
}
\description{
  Gives the "basenames" (i.e. names undecorated by the id
  string) of the parameters.
}
\seealso{
  \code{\link{CovarianceSE}}
}

