\name{getParamNamesPlain.CovarianceSEAniso2D}
\alias{CovarianceSEAniso2D$paramNamesPlain}
\alias{getParamNamesPlain.CovarianceSEAniso2D}
\title{Basenames of parameters}
\arguments{
  \item{...}{Not used.}
}
\value{
  The basenames of the parameters.
}
\description{
  Gives the "basenames" (i.e. names undecorated by the id
  string) of the parameters.
}
\seealso{
  \code{\link{CovarianceSEAniso2D}}
}

