\name{getParamNamesPlain.CovarianceSEVaryingEll}
\alias{CovarianceSEVaryingEll$paramNamesPlain}
\alias{getParamNamesPlain.CovarianceSEVaryingEll}
\title{Basenames of parameters}
\arguments{
  \item{...}{Not used.}
}
\value{
  The basenames of the parameters.
}
\description{
  Gives the "basenames" (i.e. names undecorated by the id
  string) of the parameters.
}
\seealso{
  \code{\link{CovarianceSEVaryingEll}}
}

