\name{getParamsPlain.CovarianceSEVaryingEll}
\alias{CovarianceSEVaryingEll$paramsPlain}
\alias{getParamsPlain.CovarianceSEVaryingEll}
\title{Parameter values with plain names}
\arguments{
  \item{...}{Not used.}
}
\value{
  The parameters for this covariance function, but with
  names undecorated by its id.
}
\description{
  Gives a vector of parameter values, whose names are NOT
  decorated by the id of this Covariance object.
}
\seealso{
  \code{\link{setParamsPlain.Covariance}}

  \code{\link{CovarianceSEVaryingEll}}
}

