\name{getVaryingParamNames.Model}
\alias{Model$varyingParamNames}
\alias{getVaryingParamNames.Model}
\title{Non-constant Model parameters}
\arguments{
  \item{this}{The Model object.}

  \item{...}{Not used.}
}
\value{
  The names of all parameters which are not constant.
}
\description{
  Names of the parameters which are not constant.
}
\seealso{
  \code{\link{Model}}
}

