\name{getX.Dataset}
\alias{getX}
\alias{getX.Dataset}
\alias{x}
\title{X-values}
\arguments{
  \item{...}{Not used.}
}
\value{
  (\code{n} x \code{d}) matrix giving the X-values for this
  Dataset.
}
\description{
  Gives the X-values of the Dataset: i.e., the covariates
  where we have observations.
}
\details{
  Usually accesed using \code{this$x} syntax; read-only.
}
\seealso{
  \code{\link{Dataset}}
}

