\name{NamedCovariance.Model}
\alias{NamedCovariance}
\alias{NamedCovariance.Model}
\title{Retrieve one contributing Covariance}
\arguments{
  \item{this}{The Model object.}

  \item{id}{(character) The ID of the Covariance to
  retrieve.}

  \item{...}{Not used.}
}
\value{
  A clone of the first contributing Covariance objects with
  the given id.
}
\description{
  Retrieve a clone of the first contributing Covariance
  objects with the given id.
}
\seealso{
  \code{\link{Model}}

  \code{\link{Covariance}}
}

