\name{PlotMatrixQuickAndDirty}
\alias{PlotMatrixQuickAndDirty}
\title{Plots a numeric matrix}
\usage{
  PlotMatrixQuickAndDirty(M, ...)
}
\arguments{
  \item{M}{The numeric matrix to plot.}

  \item{...}{Unused}
}
\description{
  Makes a quick plot of numeric matrix data.  Not very
  customizable, but handy for a first glance.
}
\note{
  Adapted from:
  \url{http://www.phaget4.org/R/image_matrix.html}
  2011-01-06 15:35
}

