%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  R/Covariance.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setId.Covariance}
\alias{setId.Covariance}
\alias{Covariance.setId}
\alias{setId.Covariance}
\alias{setId,Covariance-method}


\title{The ID for this Covariance}

\description{
  The ID for this Covariance:
  a character string identifying this Covariance object.
}

\usage{\method{setId}{Covariance}(this, id, ...)}

\arguments{
  \item{id}{A character string identifying this Covariance object.}
}

\seealso{
  \code{\link[R:getId.Covariance]{*getId}()}
  For more information see \code{\link{Covariance}}.
}

\author{Charles R. Hogg III}
\keyword{internal}
\keyword{methods}
