\name{StoreMatrix.LazyMatrix}
\alias{StoreMatrix}
\alias{StoreMatrix.LazyMatrix}
\title{Stores a calculated matrix}
\arguments{
  \item{M}{The matrix to store.}

  \item{...}{Not used.}
}
\description{
  Stores the supplied matrix in memory, and remembers the
  ingredients used to calculate it.
}
\seealso{
  \code{\link{LazyMatrix}}
}

