\name{Variance.CovarianceNoise}
\alias{Variance.CovarianceNoise}
\title{Noise variance at each point}
\arguments{
  \item{X}{The points we want to know the noise variance
  at.}

  \item{...}{Not used.}
}
\value{
  A numeric vector of the same length as X, with the
  corresponding noise variance.
}
\description{
  Calculate the Noise variance of the points at X.
}
\seealso{
  \code{\link{CovarianceNoise}}
}

