\name{Variance.CovarianceSE}
\alias{Variance.CovarianceSE}
\title{SE variance at each point}
\arguments{
  \item{X}{The points we want to know the SE variance at.}

  \item{...}{Not used.}
}
\value{
  A numeric vector of the same length as X, with the
  corresponding SE variance.
}
\description{
  Calculate the SE variance of the points at X: i.e., the a
  priori uncertainty at each point.
}
\seealso{
  \code{\link{CovarianceSE}}
}

