\name{Variance.CovarianceSEAniso2D}
\alias{Variance.CovarianceSEAniso2D}
\title{Anisotropic 2D SE variance at each point}
\arguments{
  \item{X}{The points we want to know the anisotropic 2D SE
  variance at.}

  \item{...}{Not used.}
}
\value{
  A numeric vector of the same length as X, with the
  corresponding anisotropic 2D SE variance.
}
\description{
  Calculate the Anisotropic 2D SE variance of the points at
  X: i.e., the a priori uncertainty at each point.
}
\seealso{
  \code{\link{CovarianceSEAniso2D}}
}

