\name{Variance.CovarianceSELocalized}
\alias{Variance.CovarianceSELocalized}
\title{Localized SE variance at each point}
\arguments{
  \item{X}{The points we want to know the localized SE
  variance at.}

  \item{...}{Not used.}
}
\value{
  A numeric vector of the same length as X, with the
  corresponding localized SE variance.
}
\description{
  Calculate the localized SE variance of the points at X:
  i.e., the a priori uncertainty at each point.
}
\seealso{
  \code{\link{CovarianceSELocalized}}
}

