\name{Variance.CovarianceSEVaryingEll}
\alias{Variance.CovarianceSEVaryingEll}
\title{Nonstationary SE variance at each point}
\arguments{
  \item{X}{The points we want to know the nonstationary SE
  variance at.}

  \item{...}{Not used.}
}
\value{
  A numeric vector of the same length as X, with the
  corresponding nonstationary SE variance.
}
\description{
  Calculate the nonstationary SE variance of the points at
  X: i.e., the a priori uncertainty at each point.
}
\seealso{
  \code{\link{CovarianceSEVaryingEll}}
}

