\name{Variance}
\alias{Variance}
\title{SE variance at each point}
\arguments{
  \item{X}{The points we want to know the SE variance at.}

  \item{\dots}{Not used.}
}
\value{
  A numeric vector of the same length as X, with the
  corresponding SE variance.
}
\description{
  "Pure virtual" method to calculate the SE variance of the
  points at X: i.e., the a priori uncertainty at each
  point.
}

