\name{clone.Covariance}
\alias{clone.Covariance}
\title{Deep-clone a Dataset}
\usage{
  \method{clone}{Covariance} (this, ...)
}
\arguments{
  \item{this}{The Covariance object to clone.}

  \item{...}{Not used.}
}
\value{
  A deep clone of the Covariance object.
}
\description{
  Deep-clone a Dataset (i.e., clones LazyMatrix objects
  too).
}
\seealso{
  \code{\link{Covariance}}
}

