%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  R/Covariance.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getId.Covariance}
\alias{getId.Covariance}
\alias{Covariance.getId}
\alias{getId.Covariance}
\alias{getId,Covariance-method}


\title{The ID for this Covariance}

\description{
  The ID for this Covariance:
  a character string identifying this Covariance object.
}

\usage{\method{getId}{Covariance}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
  The id of this Covariance object.
}

\seealso{
  \code{\link[R:setId.Covariance]{*setId}()}
  For more information see \code{\link{Covariance}}.
}

\alias{id.Covariance}

\author{Charles R. Hogg III}
\keyword{internal}
\keyword{methods}
