\name{getId.Model}
\alias{getId.Model}
\alias{Model$id}
\alias{setId.Model}
\title{ID string for this Model}
\arguments{
  \item{this}{The \code{Model} whose contributions to
  list.}

  \item{id}{(character) The new ID for \code{this}.}

  \item{...}{Not used.}
}
\value{
  The id of \code{this}.
}
\description{
  A character string identifying this Model object.
}
\seealso{
  \code{\link{Model}}
}

