\name{getM.LazyMatrix}
\alias{getM}
\alias{getM.LazyMatrix}
\alias{M}
\title{The previously-computed matrix}
\arguments{
  \item{...}{Not used.}
}
\value{
  The previously-computed matrix.
}
\description{
  Accesses the matrix which was previously computed.
}
\seealso{
  \code{\link{LazyMatrix}}
}

