%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  R/Covariance.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getParamNames.Covariance}
\alias{getParamNames.Covariance}
\alias{Covariance.getParamNames}
\alias{getParamNames.Covariance}
\alias{getParamNames,Covariance-method}


\title{Decorated names of the covariance parameters}

\description{
  Decorated names of the covariance parameters.
}

\usage{\method{getParamNames}{Covariance}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
  Parameter names in the form "id.basename", where basename is the
  "undecorated" parameter name.
}

\seealso{
  \code{\link[R:getParams.Covariance]{*getParams}()}
  For more information see \code{\link{Covariance}}.
}

\author{Charles R. Hogg III}
\keyword{internal}
\keyword{methods}
