\name{print.Covariance}
\alias{print.Covariance}
\title{Pretty-printing for Covariance objects}
\usage{
  \method{print}{Covariance} (this, indent = 0, ...)
}
\arguments{
  \item{this}{The Covariance object to print.}

  \item{...}{Not used.}
}
\description{
  Prints out the id of the Covariance, followed by a list
  of its parameters. For each parameter, we print the name,
  lower bound, current value, and upper bound.
}
\seealso{
  \code{\link{Covariance}}
}

