%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  R/Covariance.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setParamsPlain.Covariance}
\alias{setParamsPlain.Covariance}
\alias{Covariance.setParamsPlain}
\alias{setParamsPlain.Covariance}
\alias{setParamsPlain,Covariance-method}


\title{Set parameters using undecorated names}

\description{
  Set parameters using undecorated names.
}

\usage{\method{setParamsPlain}{Covariance}(this, p, ...)}

\arguments{
  \item{p}{A (named) vector of new parameter values (we \emph{only} use ones
     which are named, and whose names match up with names of parameters.)}
  \item{...}{Not used.}
}

\seealso{
  \code{\link[R:getParamsPlain.Covariance]{*getParamsPlain}()}
  For more information see \code{\link{Covariance}}.
}

\author{Charles R. Hogg III}
\keyword{internal}
\keyword{methods}
